+++ src/components/UsernameForm/CompletionMessage.js
@@ -1,10 +1,10 @@
 import React from 'react';
-import { getQueryYear as year } from '../../utils/utils';
+import { getYear } from '../../utils/date';
 
 const CompletionMessage = () => {
   return (
     <div className="text-center text-green-dark pb-2">
-      <p className="text-xl">You have successfully completed Frogtoberfest {year()}!</p>
+      <p className="text-xl">You have successfully completed Frogtoberfest {getYear()}!</p>
     </div>
   );
 };
