+++ src/pages/User/components/PullRequests/PullRequest/PullRequestInfo.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { formatDate } from '../../../../../utils/utils';
+import { formatDate } from '../../../../../utils/date';
 
 const PullRequestInfo = ({ pullRequest }) => (
   <div>
