+++ src/pages/User/components/PullRequests/index.js
@@ -7,9 +7,8 @@ import IssuesLink from './IssuesLink';
 import MeLinkInfo from './MeLinkInfo';
 import ErrorText from './ErrorText';
 import UserInfo from './UserInfo';
-import { fetchInfoFromGitHub } from '../../../../utils/utils';
-import { GITHUB_TOKEN, TOTAL_PR_COUNT, TOTAL_OTHER_PR_COUNT, GITHUB_ORG_NAME, LF_CAREER_URL } from '../../../../config';
-import { getQueryYear } from '../../../../utils/utils';
+import { fetchInfoFromGitHub, getApiUrls } from '../../../../services/index';
+import { GITHUB_TOKEN, TOTAL_PR_COUNT, TOTAL_OTHER_PR_COUNT, LF_CAREER_URL } from '../../../../config';
 
 /**
  * Returns an object containing user info.
@@ -18,13 +17,8 @@ import { getQueryYear } from '../../../../utils/utils';
  * @returns {*}
  */
 export async function fetchUserInfo(username) {
-  const queryYear = getQueryYear();
+  const apiUrls = getApiUrls(username);
 
-  const apiUrls = [
-    `https://api.github.com/search/issues?q=author:${username}+is:pr+created:${queryYear}-10-01..${queryYear}-10-30`,
-    `https://api.github.com/search/users?q=user:${username}`,
-    `https://api.github.com/orgs/${GITHUB_ORG_NAME}/members/${username}`
-  ];
   const results = apiUrls.map(url => fetchInfoFromGitHub(url, GITHUB_TOKEN));
   let [data, userDetail, membershipStatus] = await Promise.all(results);
